Sample Query

+++ Models/Book.cs
@@ -6,6 +6,6 @@ public class Book
     public string Name { get; set; }
     public float Price { get; set; }
     public string AuthorName { get; set; }
-    public int WriterId { get; set; }
+    public long WriterId { get; set; }
     public List<Genre> Genres { get; set; } = new();
 }
